fix: hammer: Rename material-didatico component to trakto-edu, add to routing and fix forward function
