+++ src/app/components/modules-page/modules-page.component.html
@@ -5,7 +5,7 @@
       moduleTitle="Material didático"
       moduleColor="#4173CD"
       moduleImage="material-didatico-module"
-      (moduleClick)="goto('/material-didatico')"
+      (moduleClick)="goto('/trakto-edu')"
     ></app-module>
     <app-module
       moduleTitle="Quiz"
