+++ src/app/components/trakto-edu-page/trakto-edu-page.component.html
@@ -1,0 +1,62 @@
+<app-header [dark]="false"></app-header>
+<div class="container">
+  <img src="../../../assets/jornada.jpg" alt="trakto-edu-main-image" class="jornada-image"/>
+  <div class="trakto-edu-content">
+    <div class="trakto-edu-header">
+      <div class="trakto-edu-text">
+        <div class="week-button-container">
+          <div class="week-container">
+            <p class="week-to-week">{{ currentWeek }}</p>
+          </div>
+          <button class="carousel-button" (click)="prevSlide.call(this)">
+            <img src="../../../assets/arrow-left.png"/>
+          </button>
+          <button class="carousel-button" (click)="nextSlide.call(this)">
+            <img src="../../../assets/arrow-right.png"/>
+          </button>
+        </div>
+        <p class="carousel-header">Material didático</p>
+        <p class="carousel-subtitle">Acesse suas últimas apresentações de aula.</p>
+      </div>
+      <button class="see-all-button" (click)="openMaterial()">Ver todo conteúdo</button>
+    </div>
+    <div class="keen-slider" #sliderRef>
+      <div *ngFor="let slide of slides" class="keen-slider__slide">
+        <button (click)="goToDocument(slide.id)" class="clickable-image">
+          <img [src]="slide.imageUrl" [alt]="slide.alt" class="carousel-image">
+          <p class="slide-title">{{ slide.title }}</p>
+        </button>
+      </div>
+    </div>
+  </div>
+  <div class="trakto-edu-content">
+    <div class="trakto-edu-header">
+      <div class="trakto-edu-text">
+        <div class="week-button-container">
+          <div class="week-container">
+            <p class="week-to-week">{{ currentWeek }}</p>
+          </div>
+          <button class="carousel-button" (click)="prevThumbnail.call(this)">
+            <img src="../../../assets/arrow-left.png"/>
+          </button>
+          <button class="carousel-button" (click)="nextThumbnail.call(this)">
+            <img src="../../../assets/arrow-right.png"/>
+          </button>
+        </div>
+        <p class="carousel-header">Vídeo aulas</p>
+        <p class="carousel-subtitle">Complemente os estudos assistindo o conteúdo de forma dinâmica.</p>
+      </div>
+      <button class="youtube-button" (click)="openTraktoChannel()">
+        <img src="../../../assets/youtube-icon.svg" alt="Youtube icon" class="youtube-icon"/>
+        Abrir no Youtube
+      </button>
+    </div>
+    <div class="keen-slider" #sliderVideo>
+      <div *ngFor="let thumbnail of thumbnails" class="keen-slider__slide">
+        <button (click)="goToVideo(thumbnail.url)" class="clickable-image">
+          <img [src]="thumbnail.src" [alt]="thumbnail.alt" class="carousel-image">
+        </button>
+      </div>
+    </div>
+  </div>
+</div>
